Refactor AIState into separate managers
